create contract transportation and update prices

+++ Core/src/main/java/ir/sharif/gamein2021/core/domain/dto/ContractSupplierDto.java
@@ -24,7 +24,7 @@ public class ContractSupplierDto
     private Integer boughtAmount;
     private Enums.VehicleType transportType;
     private boolean hasInsurance;
-    private Integer transportationCost;
+    private Float transportationCost;
     private Integer terminatePenalty;
     private boolean isTerminated;
     private Integer noMoneyPenalty;
